SM-b5f46
Moved the sites fetching to the sites store.
Removed isLoading from the useRequest composable.
Other small refactors.
