+++ client/src/stores/sites.store.ts
@@ -1,6 +1,7 @@
 import { defineStore } from "pinia";
-import { TSite } from "../types.ts";
+import { TSite, TSitesResponse } from "../types.ts";
 import { ref } from "vue";
+import { useRequest } from "../composables";
 
 export const useSitesStore = defineStore('sites', () => {
 
@@ -15,5 +16,22 @@ export const useSitesStore = defineStore('sites', () => {
     lastFetch.value = Date.now();
   }
 
-  return { sites, lastFetch, addSite, setLastFetch };
+  const fetchSites = async (): Promise<void | Error> => {
+    if (Date.now() - lastFetch.value < 30000) return;
+
+    const { data, error } = await useRequest<TSitesResponse>({
+      endpoint: '/sites/all',
+      needsAuth: true,
+    });
+
+    if (error) return error;
+
+    if (data && data.sites) {
+      sites.value = [...data.sites];
+    }
+
+    setLastFetch();
+  }
+
+  return { sites, lastFetch, addSite, fetchSites, setLastFetch };
 });
