+++ client/vite.config.ts
@@ -7,4 +7,4 @@ export default defineConfig({
     outDir: './dist'
   },
   plugins: [vue()],
-})
+});
